6/20

+++ engine/config/config_value.h
@@ -6,6 +6,6 @@
 using ConfigValue = std::variant<
     bool,
     int,
-    float,
+    double,
     std::string
 >;
